✨ (page.tsx): Update project information and add new projects to the list The commit introduces new projects to the list and updates existing project information. New projects such as EMP Consultants, Top Timeleines, BEAUTY By Casey Lauren, Taoist Programmer, Taoist Shop, Slicks Slices, Slick Fits, and Dang That's Delicious are added with their respective descriptions, links, and logos. The project information is updated to reflect the changes accurately.
✨ (page.tsx): Add a new page component for displaying technology stack information
ℹ️ (page.tsx): Include detailed information about various technologies such as JavaScript, TypeScript, Node.js, Express.js, React, Next.js, Vite, Remix, Gatsby, NPM, Yarn, GraphQL, React Query, React Router, Tailwind CSS, Tailwind UI, Headless UI, Keystone.js, MongoDB Atlas, Supabase, Netlify, Digital Ocean, and Vercel
The changes introduce a new page component to display detailed information about various technologies used in the project. This addition aims to provide users with comprehensive insights into the technology stack utilized, enhancing transparency and understanding of the project's architecture and dependencies.

✨ (TechStack): Add metadata and update TechStack component to display a list of technologies with logos and descriptions
The changes include adding metadata for the Tech Stack page with a title and description. The TechStack component has been updated to display a list of technologies with logos, names, descriptions, and links. This enhancement improves the user experience by providing more information about the technologies used in the project.

📝 (page.tsx): Update tools and resources in the Uses component
The page.tsx file has been updated to reflect the latest tools and resources used by the developer. The changes include updating the workstation tools with new hardware specifications, adding new development tools like Visual Studio Code and GitKraken, and including new design tools like Whimsical and Pixelmator Pro. Additionally, new podcasts have been added to the list of recommended podcasts for developers. These updates aim to provide more relevant and up-to-date information to the users of the application.